# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "",
        "version": "1.0.6",
        "title": "Swagger Links",
        "contact": {
            "email": "luis.alvarez@bewe.io"
        }
        
    },
    "host": "http://ec2-34-224-5-243.compute-1.amazonaws.com:3000",
    "basePath": "/v2",
    "schemes": [
        "http"
    ],
    "paths": {
        "/links/add": {
            "post": {
                "tags": [
                    "link"
                ],
                "summary": "Creates links of user",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created link object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Link"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful"
                    }
                }
            }
        },
        "/links": {
            "get": {
                "tags": [
                    "link"
                ],
                "summary": "Get link by user",
                "description": "",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "schema": {
                            "$ref": "#/definitions/LinkData"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "link"
                ],
                "summary": "Delete link",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteLink",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id ",
                        "in": "path",
                        "description": "The id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error message"
                    },
                    "200":{
                        "description":"success"
                    } 
                }               
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Logs user into the system",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user email for login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "headers": {
                            "X-Expires-After": {
                                "type": "string",
                                "format": "date-time",
                                "description": "date in UTC when token expires"
                            },
                            "X-Rate-Limit": {
                                "type": "integer",
                                "format": "int32",
                                "description": "calls per hour allowed by the user"
                            }
                        },
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid credentials"
                    }
                }
            }
        },
        "/auth/singin": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "createUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        },
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
            "flow": "implicit",
            "scopes": {
                "read:pets": "read your pets",
                "write:pets": "modify pets in your account"
            }
        }
    },
    "definitions": {
        "LinkData":{
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/Link"
                }
            },
            "xml": {
                "name": "Category"
            }
        },
        "Link": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Category"
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}